feat: guards for if user is logged in
Depending whether the user is logged in or not will make the view act
differently. If the user is logged in and when they attempt to register
for the bracket (assuming that the conditions for the bracket are
sufficient), the user will then be taken to the bracket form view.

Else, the user will be redirected to the login page