fix: improved on InsertGamerToBracket
The method now first checks the bracket that the bracket player is being
inserted to. If the player count of the bracket is 128 and over, then
the insertion will fail. Else, the bracket player will be inserted into
the bracket_player_info table and the number of players in that bracket
will be updated